functions/requirements.txt
functions-framework==3.*
a2wsgi==1.10.0
firebase-admin>=6.5.0
firebase-functions>=0.1.0
fastapi==0.128.0
uvicorn==0.40.0
python-dotenv>=1.0.1
pydantic==2.12.5
google-cloud-firestore>=2.11.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator
beautifulsoup4==4.12.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
greenlet==3.3.1
h11==0.16.0
idna==3.11
pydantic_core==2.41.5
requests==2.32.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
python-multipart==0.0.9
